+++ scripts/macros/monsterFeatures/generic/autoPush.js
@@ -5,7 +5,12 @@ async function use({trigger, workflow}) {
     if (isNaN(Number(config.distance))) return;
     workflow.targets.forEach(token => {
         if (config.failed && !workflow.failedSaves.has(token)) return;
-        tokenUtils.pushToken(workflow.token, token, Number(config.distance));
+        let distance = Number(config.distance);
+        if (distance < 0) {
+            let distanceBetween = tokenUtils.getDistance(workflow.token, token);
+            distance = Math.max(distance, -distanceBetween);
+        }
+        tokenUtils.pushToken(workflow.token, token, distance);
     });
 }
 export let autoPush = {